move interface and functionality for fetching gamedetails to GameUtils
